refactor: move register server action to lib
